minor modification on x axis for one histo

+++ DQM/TrackingMonitor/python/TrackingMonitorSeedNumber_cff.py
@@ -12,7 +12,7 @@ TrackMonStep0.TCProducer    = cms.InputTag("initialStepTrackCandidates")
 TrackMonStep0.AlgoName      = cms.string('iter0')
 
 TrackMonStep0.TkSeedSizeBin = cms.int32(100)
-TrackMonStep0.TkSeedSizeMax = cms.double(5000)                         
+TrackMonStep0.TkSeedSizeMax = cms.double(20000)                         
 TrackMonStep0.TkSeedSizeMin = cms.double(0)
 TrackMonStep0.NClusPxBin = cms.int32(100)
 TrackMonStep0.NClusPxMax = cms.double(5000)
